+++ src/main/java/com/lhy/createpackage/registry/ModCreativeTabs.java
@@ -23,6 +23,7 @@ public final class ModCreativeTabs {
                     .icon(() -> new ItemStack(ModItems.PACKAGE_DISTRIBUTOR.get()))
                     .displayItems((parameters, output) -> {
                         output.accept(ModItems.PACKAGE_DISTRIBUTOR.get());
+                        output.accept(ModItems.ADVANCED_PACKAGE_DISTRIBUTOR.get());
                         output.accept(ModItems.MACHINE_LINKER.get());
                     })
                     .build());